deprecated Pet.py
updated old Pet.py, moved so all pet-related functions/assets are in the "Pet" folder
